neue email Categorie eingefügt

+++ NSString+MD5.m
@@ -13,7 +13,7 @@
 - (NSString *)MD5String {
     const char *cstr = [self UTF8String];
     unsigned char result[16];
-    CC_MD5(cstr, strlen(cstr), result);
+    CC_MD5(cstr, (unsigned int)strlen(cstr), result);
     
     return [NSString stringWithFormat:
             @"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
